fixed bugs from presentation
Bug Fix: Lines not showing for reflections of nothing, the issue was that when calling trace on the scene a second time I was passing in this instead of camera.

Bug Fix: lights in scene 0 and 8 not being yellow. Fix was to add an alpha = 1.0 in light. I guess that for those scenes it for some reason reduced the opacity, but nothing is drawn before it to make it opaque.